Change birth spacing preference for 0 months postpartum
This was historically at 0.8 at PP mont h0, higher than the other immediate PP months.  With other tweaking that has happened since then (combination of many) we no longer need these increased very-short-interval births.  If we are going to display birth spacing preference as a curve we should get rid of this odd shape as there is no need or reason for it.


Former-commit-id: a326816e9029235667aabd853df7e7737f4cc19c